fix(web-frontend/shift-type-requirements): Update `preferred_num_people` when `required_num_people` has been changed
If required_num_people has been changed to same as preferred_num_people, change preferred_num_people to undefined.

Before this commit, consider a scenario with `required = 12` and `preferred = 13`. If we updated `required` to 13, `preferred` remained at 13 instead of being reset to `undefined` as intended. This commit fixes that issue.